+++ bench/nested-deps-app-router-many-pages/template/layout.js
@@ -1,14 +1,5 @@
 import React from 'react'
 
-export const metadata = {
-  title: 'Next.js',
-  description: 'Generated by Next.js',
-}
-
-export default function RootLayout({ children }) {
-  return (
-    <html lang="en">
-      <body>{children}</body>
-    </html>
-  )
+export default function Layout({ children }) {
+  return <main>{children}</main>
 }
