+++ bench/nested-deps-app-router-many-pages/template/root-layout.js
@@ -1,0 +1,14 @@
+import React from 'react'
+
+export const metadata = {
+  title: 'Next.js',
+  description: 'Generated by Next.js',
+}
+
+export default function RootLayout({ children }) {
+  return (
+    <html lang="en">
+      <body>{children}</body>
+    </html>
+  )
+}
